Lab 8 Gate

+++ app/Http/Controllers/ArticleController.php
@@ -28,6 +28,7 @@ class ArticleController extends Controller
      */
     public function create()
     {
+        $this->authorize('create', [self::class]);
         return view('articles.create');
     }
     /**
@@ -71,7 +72,8 @@ class ArticleController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function edit($id)
-    {
+    {   
+        $this->authorize('update', [self::class]);
         $article = Article::FindOrFail($id);
         return view('articles.edit', ['article' => $article]);
     }
@@ -106,6 +108,7 @@ class ArticleController extends Controller
      */
     public function destroy($id)
     {
+        $this->authorize('delete', [self::class]);
         $article = Article::FindOrFail($id);
         Comment::where('article_id', $id)->delete();
         $article->delete();
